refactor(api): implementa roteamento modular
- Move roteadores para dentro de seus respectivos módulos (`morador`, `funcionario`, `chave`, `autenticacao`, `core`).
- Transforma cada diretório de módulo em um componente autossuficiente (Router -> Controller -> Service -> Model).
- Simplifica o agregador central de rotas em `api/app/routers/__init__.py`.
- Melhora a localidade do código e facilita a manutenção por domínio.
- Garante conformidade total com Ruff e passa em todos os 11 testes integrados.
